fifo driver completely working, all tests pass
The fifo driver is completely working and all tests
pass.  It has been tested with a FIFO size up to 8192 bytes.

Fixed a bug with the buffer memory.
Previously the buffer was allocated as an array in the
per-device structure in init.  Inside read/write it
was attempted to get a pointer to this array and step
through the buffer using pointer arithmetic.
Unfortunately the pointer only pointed to the first value.
To fix this the buffer was changed to a pointer (not an array)
and was properly kmalloc'ed.