fixed a possible bug involving size_t and loff_t
Previously the file offset was stored as an unsigned type (size_t).
However the offset should support negative values (loff_t).
For example SEEK_SET with -1 would have resulted in a very large
integer.  While this still would have been caught it was being
caught for being too large when it should have been because it was
too small.