changed doc .pdf pointer to doc-build branch
Instead of including the built doc .pdf in the master
branch, which makes it difficult to delete, it has
been moved to its own branch (doc-build) which can
be easily deleted.  Binary blobs like this can be
convenient to have around but they can consume a
lot of space.  The diffs of these files are often
useless as well.