added misc char driver example
A driver has been added that uses the misc char operations.  This
simplifies the creation of sysfs and other entries for simple char
drivers.

This driver stores a unique id.  It can be read out.  Data can also be
written.  If the data written is not the unique id an error is returned.

This driver is a good example of how to use the loff_t to keep track of
the offset during read/write.